+++ tests/monkeypatch.py
@@ -63,11 +63,12 @@ def _monkeypatch_psutil(monkeypatch: pytest.MonkeyPatch) -> None:
 
 
 def _monkeypatch_docker(monkeypatch: pytest.MonkeyPatch) -> None:
-    class FakeDockerClient:
-        def ping(self: FakeDockerClient) -> bool:
-            return False
+    from ubo_app.utils.fake import Fake
 
-    monkeypatch.setattr('docker.from_env', lambda: FakeDockerClient())
+    monkeypatch.setattr(
+        'docker.from_env',
+        lambda: Fake(_Fake__props={'ping': Fake(_Fake__return_value=False)}),
+    )
 
 
 def _monkeypatch_datetime(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -188,9 +189,9 @@ def _monkeypatch_subprocess(monkeypatch: pytest.MonkeyPatch) -> None:
             ):
                 return original_subprocess_run(command, *cast(Any, args), **kwargs)
             # Reboot and poweroff
-            if command[1] == 'systemctl' and command[2] in ('reboot', 'poweroff'):
+            if command[1] == 'systemctl' and command[2] in {'reboot', 'poweroff'}:
                 return Fake()
-        if command[0] in ('cat', 'file'):
+        if command[0] in {'cat', 'file'}:
             return original_subprocess_run(command, *cast(Any, args), **kwargs)
         msg = f'Unexpected `subprocess.run` command in test environment: {command}'
         raise ValueError(msg)
@@ -260,7 +261,6 @@ def _monkeypatch(monkeypatch: pytest.MonkeyPatch) -> None:
     import importlib.metadata
 
     import ubo_app.constants
-    import ubo_app.store.services.notifications
     import ubo_app.utils.serializer
     from ubo_app.utils.fake import Fake
 
@@ -270,7 +270,7 @@ def _monkeypatch(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(importlib.metadata, 'version', lambda _: '0.0.0')
 
     monkeypatch.setattr(ubo_app.constants, 'STORE_GRACE_PERIOD', 0.1)
-    monkeypatch.setattr(ubo_app.store.services.notifications, 'FLASH_TIME', 1000)
+    monkeypatch.setattr(ubo_app.constants, 'NOTIFICATIONS_FLASH_TIME', 1000)
     monkeypatch.setattr(ubo_app.utils.serializer, 'add_type_field', lambda _, obj: obj)
 
     sys.modules['ubo_app.utils.secrets'] = Fake(