+++ ubo_app/utils/hardware.py
@@ -1,5 +1,6 @@
 # pyright: reportMissingModuleSource=false
 # ruff: noqa: D100, D101, D102, D103, D104, D107
+
 from ubo_app.utils import IS_RPI
 
 
@@ -11,23 +12,3 @@ def initialize_board() -> None:
 
     GPIO.setmode(GPIO.BCM)
     GPIO.setup(17, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
-
-
-def turn_off_screen() -> None:
-    if not IS_RPI:
-        return
-    from RPi import GPIO
-
-    GPIO.setup(26, GPIO.OUT)
-    GPIO.output(26, GPIO.LOW)
-
-
-def turn_on_screen() -> None:
-    if not IS_RPI:
-        return
-
-    from RPi import GPIO
-
-    GPIO.setmode(GPIO.BCM)
-    GPIO.setup(26, GPIO.OUT)
-    GPIO.output(26, GPIO.HIGH)